add Database

See merge request Antoska/tfs-android-fall-2021!7

+++ TinkoffMessenger/app/src/main/java/ru/myacademyhomework/tinkoffmessenger/ChatMessageListener.kt
@@ -1,5 +1,5 @@
 package ru.myacademyhomework.tinkoffmessenger
 
 interface ChatMessageListener {
-    fun itemLongClicked(idMessage: Int, position: Int): Boolean
+    fun itemLongClicked(idMessage: Long, position: Int): Boolean
 }
